Add cppcheck integration and fix printf format specifiers
- Add comprehensive test system to CMakeLists.txt with 31 test targets
- Integrate cppcheck static analysis with multiple targets:
  - 'cppcheck' for quick analysis
  - 'cppcheck-full' for comprehensive analysis
  - 'cppcheck-xml' for CI integration
  - 'lint' combined target
  - 'quickcheck' for fast validation
- Add enable_testing() at root level for proper ctest integration
- Fix printf format specifiers using stdint.h macros (PRIu32, PRId32, etc.):
  - databox.c: Use PRIu32 for databoxType comparisons
  - dataspeed.c: Use %zu for size_t
  - dks.c: Use PRIu32 for enum type comparisons
  - flex.c: Use PRIdPTR for ptrdiff_t, PRIu32 for uint32_t
  - hyperloglog.c: Use PRIu32/PRId32/PRIu8 for proper type formatting
  - multiroar.c: Use PRIu16/PRIu32/PRIu64/%zu for various types
- Fix missing return statements in flex.c after flexFind__ macro calls

+++ src/databox.c
@@ -870,7 +870,7 @@ int databoxCompare(const databox *restrict a, const databox *restrict b) {
             return -databoxCompareInt128Float(b, a);
         default:
 #ifndef NDEBUG
-            printf("Attempted to compare: %d and %d?", a->type, b->type);
+            printf("Attempted to compare: %" PRIu32 " and %" PRIu32 "?", (uint32_t)a->type, (uint32_t)b->type);
 #endif
             assert(NULL && "Either you tried to compare against DATABOX_VOID "
                            "or we created new numeric types and didn't add "
